refactor :hammer: refactor SearchLocationController with dataSource followed by MVC Pattern
