--- iOS/JacksonBnB/JacksonBnB/Controller/ResultHotelsViewController.swift
@@ -1,9 +1,0 @@
-//
-//  ResultHotelsViewController.swift
-//  JacksonBnB
-//
-//  Created by jinseo park on 5/21/21.
-//
-
-import Foundation
-//안쓸듯
