+++ iOS/JacksonBnB/JacksonBnB/Controller/SearchLocationsViewController.swift
@@ -7,58 +7,33 @@
 
 import UIKit
 
-class SearchLocationsViewController: UIViewController, UICollectionViewDataSource, UICollectionViewDelegate, UISearchResultsUpdating {
+class SearchLocationsViewController: UIViewController {
     
-    let searchLocationsController = UISearchController()
-    
-    let dbManager = DataBaseManager()
-    var allLocations = Locations(locations: [])
-    var filteredLocations = Locations(locations: [])
-    
-    @IBOutlet weak var collectionView: UICollectionView!
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        if searchLocationsController.isActive {
-            return filteredLocations.locations.count
-        }else {
-            return allLocations.locations.count
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: LocationCell.reuseIdentifier, for: indexPath) as! LocationCell
-        if searchLocationsController.isActive {
-            cell.locationNameLabel.text = filteredLocations.locations[indexPath.row].name
-            cell.locationCellImageView.image = UIImage(named: filteredLocations.locations[indexPath.row].imageName)
-        }else {
-            cell.locationNameLabel.text = allLocations.locations[indexPath.row].name
-            cell.locationCellImageView.image = UIImage(named: allLocations.locations[indexPath.row].imageName)
-        }
-        return cell
-    }
-    
-  
+    var searchLocationDataSource: SearchLocationDataSource?
     @IBOutlet weak var locationsCollectionView: UICollectionView!
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-        locationsCollectionView.delegate = self
-        locationsCollectionView.dataSource = self
-        
-        let locationNib = UINib(nibName: LocationCell.reuseIdentifier, bundle: nil)
-        locationsCollectionView.register(locationNib, forCellWithReuseIdentifier: LocationCell.reuseIdentifier)
-                        
-        locationsCollectionView.collectionViewLayout = setCollectionViewLayout()
-        allLocations = dbManager.getLocations()
+        searchLocationDataSource = SearchLocationDataSource()
+        self.locationsCollectionView.dataSource = searchLocationDataSource
         title = "숙소 찾기"
-        
-        searchLocationsController.searchResultsUpdater = self
-        
+                
+        locationsCollectionView.collectionViewLayout = setCollectionViewLayout()
+        setLocationNib()
+        setNotificationCenter()
+        setSearchBarContoller()
         
     }
-    override func viewDidAppear(_ animated: Bool) {
-
-        self.navigationItem.searchController = searchLocationsController
-        self.navigationItem.hidesSearchBarWhenScrolling = false
+    
+    func setLocationNib() {
+        let locationNib = UINib(nibName: LocationCell.reuseIdentifier, bundle: nil)
         
+        locationsCollectionView.register(locationNib, forCellWithReuseIdentifier: LocationCell.reuseIdentifier)
+    }
+    
+    func setSearchBarContoller() {
+        self.navigationItem.searchController = searchLocationDataSource?.searchLocationsController
+        self.navigationItem.hidesSearchBarWhenScrolling = false
     }
     
     func setCollectionViewLayout() -> UICollectionViewLayout { //콜렉션 뷰 레이아웃 설정
@@ -71,11 +46,14 @@ class SearchLocationsViewController: UIViewController, UICollectionViewDataSourc
         return layout
     }
     
-    func updateSearchResults(for searchController: UISearchController) {
-        guard let text = self.searchLocationsController.searchBar.text else {
-            return
-        }    
-        filteredLocations = dbManager.getFilteredLocations(by: text)
-        self.collectionView.reloadData()
+    func setNotificationCenter() {
+        NotificationCenter.default.addObserver(self, selector: #selector (SearchLocationsViewController.changeCollectionView), name: Notification.Name("cellsChanged"), object: nil)
     }
+    
+    @objc func changeCollectionView() {
+        DispatchQueue.main.async {
+            self.locationsCollectionView.reloadData()
+        }
+    }
+    
 }
